Remove Sealights checks in release-v0.7 branch
My thinking is that the value of Sealights is mainly about coverage
tracking, which is useful in main branch but less useful in a
release branch, since there we mainly get dependency bumps and
Konflux ref updates.

Also Sealights has been a little unstable lately, which I'd rather
not have to worry about when trying to get a fresh release branch
build shipped to customers.

Note that we still have all the same test coverage in the
existing checks-codecov workflow.

Ref: https://issues.redhat.com/browse/EC-1384