Update TAS version in CPE label
Because Conforma v0.7 is (or will be) released in RHTAS 1.3, let's
use the more accurate version in the CPE.

This is consistent with the change I'm making in the v0.5 and v0.6
branch.

See also PR #2950 where the CPE label was added. See also PR #2999
and #3000 which contain related v0.5 and v0.6 changes.