Followup acceptance test snapshot updates
I don't want to squash this with the previous commit, even though
it's very closely related. Why? Putting these together makes
cherry-picking to another branch more difficult. Actually it doesn't
matter now, but I'm trying to set a precent for when we do it all
again in six months.

Changes made like this:
  UPDATE_SNAPS=true make acceptance

(Actually I might cherry-pick the parent commit to the v0.7 branch
shortly, which means it is useful right now.)